По ссылке на ФИО пациента в новом окне открывается страница пациента, которая по сути представляет собой ту же patient.jsp, но кнопка редактирования личных данных и кнопка заполнения анкет недоступны. Исправила UserService и UserRepo. Id в классе у нас Integer, а в репо и сервисе был Long

+++ src/main/java/ru/careportal/core/data/UserRepo.java
@@ -5,6 +5,6 @@ import ru.careportal.core.db.model.User;
 
 import java.util.Optional;
 
-public interface UserRepo extends CrudRepository<User, Long> {
+public interface UserRepo extends CrudRepository<User, Integer> {
     Optional<User> findByEmail(String email);
 }
